Median of two sorted arrays of different size
Given two sorted arrays a and b of size m and n respectively. You need to find and return the median of the two sorted arrays.

Note: Assume that the two sorted arrays are merged and then median is selected from the combined array.

Examples:

Input: a[] = [1, 5, 9], b[] = [2, 3, 6, 7]
Output: 5
Explanation: If we merge the arrays and sort them then it will become [1, 2, 3, 5, 6, 7, 9] The middle element for [1, 2, 3, 5, 6, 7, 9] is 5
Input: a[] = [4, 6], b[] = [1, 2, 3, 5]
Output: 3.5
Explanation: If we merge the arrays and sort them then it will become [1, 2, 3, 4, 5, 6]. Since the arrays size of now even, hence the median is the average of the two middle elements.
The two middle elements here are 3 and 4 hence the median is (3+4)/2 = 3.5
Input: a[] = [4], b[] = [1, 2]
Output: 2
Explanation: If we merge the arrays then it will become [1, 2, 4]. The median of the resultant array is the middle element after sorting it. Hence the answer is 2.