Sorted insert for circular linked list
Given a sorted circular linked list, the task is to insert a new node in this circular list so that it remains a sorted circular linked list.

Examples:

Input: LinkedList: 1->2->4 (the first and last node is connected, i.e. 4 -> 1), data = 2
Output: 1->2->2->4
Explanation: We can add 2 after the second node.

Input: LinkedList = 1->4->7->9 (the first and last node is connected, i.e. 9 --> 1), data = 5
Output: 1->4->5->7->9
Explanation: We can add 5 after the second node.