Longest Palindrome in a String
Given a string s, your task is to find the longest palindromic substring within s.

A substring is a contiguous sequence of characters within a string, defined as s[i...j] where 0 ≤ i ≤ j < len(s).

A palindrome is a string that reads the same forward and backward. More formally, s is a palindrome if reverse(s) == s.

Note: If there are multiple palindromic substrings with the same length, return the first occurrence of the longest palindromic substring from left to right.

Examples :

Input: s= "aaaabbaa"
Output: "aabbaa"
Explanation: The longest palindromic substring is "aabbaa". 
Input: s = "abc"
Output: "a"
Explanation: "a", "b", and "c" are all palindromes of the same length, but "a" appears first. 
Input: s = "abacdfgdcaba" 
Output: "aba" 
Explanation: The longest palindromic substring is "aba", which occurs twice. The first occurrence is returned.